cilium: Remove health ep from ep manager on shutdown
Currently on shutdown we don't remove the health endpoint from the
endpoint manager but do remove the corresponding netlink interface.
Note that validateEndpoint() special cases the health endpoint and
does remove the old state directory if it existed before. But
independent of that it would be better to just streamline cleanup
via cleanupHealthEndpoint() for all cases.

Related: #32689
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Co-developed-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>